fix(terminal): copy-on-select writes nothing when WKWebView lacks focus (#827)
"选中后执行 - 复制到剪贴板" used navigator.clipboard.writeText directly,
which silently fails in WKWebView when the webview isn't first responder
— exactly the state a mouse drag selection leaves on macOS (reporter on
Tahoe 26). The right-click menu path already solved this with Wails'
Clipboard.SetText plus a browser fallback; that writer lived privately
inside useTerminalMenu.

Extract it as composables/useClipboardWrite and route all three terminal
copy paths through it: selection-action copy, macOS Cmd+C, and the
Ctrl+Shift+C event handler. Clipboard.SetText resolves false (not reject)
on glitch, so the fallback ordering is preserved.

Closes #827
